+++ src/interfaces/queue-options.ts
@@ -1,5 +1,5 @@
 import { AdvancedRepeatOptions } from './advanced-options';
-import { BaseJobOptions } from './jobs-options';
+import { DefaultJobOptions } from './base-job-options';
 import { ConnectionOptions } from './redis-options';
 
 export enum ClientType {
@@ -36,7 +36,7 @@ export interface QueueBaseOptions {
  * Options for the Queue class.
  */
 export interface QueueOptions extends QueueBaseOptions {
-  defaultJobOptions?: BaseJobOptions;
+  defaultJobOptions?: DefaultJobOptions;
 
   /**
    * Options for the rate limiter.