+++ tests/test_repeat.ts
@@ -12,7 +12,7 @@ import {
   getNextMillis,
   Worker,
 } from '../src/classes';
-import { JobsOptions } from '../src/interfaces';
+import { JobsOptions } from '../src/types';
 import { removeAllQueueData } from '../src/utils';
 
 const moment = require('moment');